Fix script not restarting due to lack of a reboot command
Forgot to add that in.

10db79d (19 Jan) states that there was a 'command not found' stderr
when using reboot. Thinking back, I don't know how that could
have happened, but I'm hoping it won't happen here, since

(1) Script is run as root in the root crontab (no permission issues)
(2) reboot is verified to be /sbin/reboot which `which reboot`,
(3) which is verified to be in the root user's $PATH (no path issues)

If error somehow shows itself again in this commit, it'll be easily
known since the stdout & stderr are logged.